+++ classes/BlueChip/Security/Modules/Log/ListTable.php
@@ -2,7 +2,8 @@
 
 namespace BlueChip\Security\Modules\Log;
 
-use BlueChip\Security\Modules\IpBlacklist;
+use BlueChip\Security\Modules\Access\Scope;
+use BlueChip\Security\Modules\InternalBlocklist\AdminPage as InternalBlocklistAdminPage;
 
 /**
  * Logs table
@@ -10,9 +11,9 @@ use BlueChip\Security\Modules\IpBlacklist;
 class ListTable extends \BlueChip\Security\Core\ListTable
 {
     /**
-     * @var string Name of blacklist action query argument
+     * @var string Name of blocklist action query argument
      */
-    private const ACTION_BLACKLIST = 'blacklist';
+    private const ACTION_BLOCKLIST = 'blocklist';
 
     /**
      * @var string Name of view query argument
@@ -253,47 +254,47 @@ class ListTable extends \BlueChip\Security\Core\ListTable
      */
     private function getRowActions(array $item): array
     {
-        if (($scope = $this->getLockScopeFromEvent($item['event'])) === IpBlacklist\LockScope::ANY) {
+        if (($scope = $this->getLockScopeFromEvent($item['event'])) === Scope::ANY) {
             // No specific scope, no action.
             return [];
         }
 
         return [
-            self::ACTION_BLACKLIST => \sprintf(
+            self::ACTION_BLOCKLIST => \sprintf(
                 '<span class="delete"><a href="%s">%s</a></span>',
                 add_query_arg(
                     [
-                        IpBlacklist\AdminPage::DEFAULT_IP_ADDRESS => $item['ip_address'],
-                        IpBlacklist\AdminPage::DEFAULT_SCOPE => $scope,
+                        InternalBlocklistAdminPage::DEFAULT_IP_ADDRESS => $item['ip_address'],
+                        InternalBlocklistAdminPage::DEFAULT_SCOPE => $scope,
                     ],
-                    IpBlacklist\AdminPage::getPageUrl()
+                    InternalBlocklistAdminPage::getPageUrl()
                 ),
-                esc_html__('Add to blacklist', 'bc-security')
+                esc_html__('Add to internal blocklist', 'bc-security')
             ),
         ];
     }
 
 
     /**
-     * Return appropriate lock scope for $event type.
+     * Return appropriate access scope for lock based on $event type.
      *
-     * @see \BlueChip\Security\Modules\IpBlacklist\LockScope
+     * @see \BlueChip\Security\Modules\Access\Scope
      *
      * @param string $event_id One from event IDs defined in \BlueChip\Security\Modules\Log\Event.
      *
-     * @return int Lock scope code. LockScope::ANY indicates that given event does not warrant blacklisting.
+     * @return int Lock scope code. Scope::ANY indicates that given event does not warrant blocklisting.
      */
     private function getLockScopeFromEvent(string $event_id): int
     {
         switch ($event_id) {
             case Events\Query404::ID:
-                return IpBlacklist\LockScope::WEBSITE;
+                return Scope::WEBSITE;
             case Events\AuthBadCookie::ID:
             case Events\LoginFailure::ID:
             case Events\LoginLockout::ID:
-                return IpBlacklist\LockScope::ADMIN;
+                return Scope::ADMIN;
             default:
-                return IpBlacklist\LockScope::ANY;
+                return Scope::ANY;
         }
     }