+++ tests/integration/src/Constants.php
@@ -8,6 +8,11 @@ namespace BlueChip\Security\Tests\Integration;
 abstract class Constants
 {
     /**
+     * @var string Username of mocked WP_User instance created automatically by WordPress test suite.
+     */
+    public const DEFAULT_USERNAME = 'admin';
+
+    /**
      * @var string Password assigned to mocked WP_User instances by WordPress test suite.
      */
     public const FACTORY_PASSWORD = 'password';
@@ -21,4 +26,14 @@ abstract class Constants
      * @var string Password that (hopefully) have not been pwned yet.
      */
     public const SAFE_PASSWORD = 'This password have not been pwned ... yet.';
+
+    /**
+     * @var string IP address that is part of Amazon Web Services cloud (us-west-1)
+     */
+    public const AMAZON_WEB_SERVICE_IP_ADDRESS = '52.93.178.234';
+
+    /**
+     * @var string IP address that is definitely not part of Amazon Web Services cloud.
+     */
+    public const CLOUDFLARE_DNS_IP_ADDRESS = '1.1.1.1';
 }